last commit

+++ app/controllers/clients_pc_controller.php
@@ -5,6 +5,7 @@
   public function index() {
      $this->title = "Clientes (Pessoa jurídica)";
      $this->clients = ClientPc::all();
+       $this->action = ViewHelpers::urlFor("/clientes/nova-pessoa-juridica");
   }
 
   public function show() {
@@ -16,9 +17,8 @@
     $this->title ="Cadastro de cliente";
     $this->cities = City::all();
     $this->client  = new ClientPc();
-    $this->action = ViewHelpers::urlFor("/clientes/pessoa-juridica");
+    $this->action = ViewHelpers::urlFor("/clientes/nova-pessoa-juridica");
     $this->submit = 'Cadastrar';
-    $this->render('new');
   }
 
   public function create(){
@@ -32,7 +32,7 @@
       $this->title ="Cadastro de cliente";
       $this->submit = 'Cadastrar';
       $this->cities = City::all();
-      $this->action = ViewHelpers::urlFor("/clientes/pessoa-juridica");
+      $this->action = ViewHelpers::urlFor("/clientes/nova-pessoa-juridica");
       $this->render('new');
     }
   }
@@ -64,8 +64,12 @@
 
   public function destroy() {
     $this->client = ClientPc::findById($this->params[':id']);
-    $this->client->deleteClient($this->params[':id']);
-    Flash::message('success', 'Cliente deletado com sucesso');
+    if(!$this->client->isTableRelations('orders','client_id')) {
+      $this->client->deleteClient($this->params[':id']);
+      Flash::message('success', 'Cliente deletado com sucesso');
+    } else {
+      Flash::message('negative', 'Cliente não pode ser deletado, pois está relacionado com outras tabelas');
+    }
     $this->redirectTo("/clientes/pessoa-juridica");
   }
